adding controllers
