+++ access.log
@@ -241,3 +241,15 @@
 ::1 - - [25/Jan/2022:00:11:33 +0000] "POST /api/verify HTTP/1.1" 404 23 "-" "PostmanRuntime/7.28.4"
 ::1 - - [25/Jan/2022:00:12:49 +0000] "POST /api/user/verify HTTP/1.1" 404 23 "-" "PostmanRuntime/7.28.4"
 ::1 - - [25/Jan/2022:00:17:17 +0000] "POST /api/users/verify HTTP/1.1" 400 50 "-" "PostmanRuntime/7.28.4"
+::1 - - [25/Jan/2022:23:31:20 +0000] "POST /api/users/login HTTP/1.1" 200 251 "-" "PostmanRuntime/7.28.4"
+::1 - - [25/Jan/2022:23:32:13 +0000] "GET /api/contacts HTTP/1.1" 200 2 "-" "PostmanRuntime/7.28.4"
+::1 - - [25/Jan/2022:23:33:06 +0000] "POST /api/contacts HTTP/1.1" 201 232 "-" "PostmanRuntime/7.28.4"
+::1 - - [25/Jan/2022:23:33:29 +0000] "GET /api/contacts HTTP/1.1" 200 234 "-" "PostmanRuntime/7.28.4"
+::1 - - [25/Jan/2022:23:42:36 +0000] "GET /api/contacts/61f088b231ba6b94ca880edb HTTP/1.1" 200 232 "-" "PostmanRuntime/7.28.4"
+::1 - - [25/Jan/2022:23:49:22 +0000] "POST /api/contacts HTTP/1.1" 201 232 "-" "PostmanRuntime/7.28.4"
+::1 - - [25/Jan/2022:23:54:32 +0000] "POST /api/contacts HTTP/1.1" 500 58 "-" "PostmanRuntime/7.28.4"
+::ffff:127.0.0.1 - - [25/Jan/2022:23:55:12 +0000] "POST /api/contacts HTTP/1.1" 201 232 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:00:03:21 +0000] "PUT /api/contacts/61f08de0690f83a1008e97a3 HTTP/1.1" 200 249 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:00:13:07 +0000] "PATCH /api/contacts/61f08de0690f83a1008e97a3/favorite HTTP/1.1" 200 250 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:00:19:34 +0000] "DELETE /api/contacts/61f08de0690f83a1008e97a3 HTTP/1.1" 200 29 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:00:20:16 +0000] "GET /api/contacts HTTP/1.1" 200 467 "-" "PostmanRuntime/7.28.4"
