+++ controllers/contacts/index.js
@@ -1,0 +1,7 @@
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const updateById = require("./updateById");
+const updateFavorite = require("./updateFavorite");
+const removeById = require("./removeById");
+module.exports = {getAll, getById, add, updateById, updateFavorite, removeById};
