+++ controllers/users/index.js
@@ -1,0 +1,2 @@
+const signup = require("./signup");
+module.exports = {signup};
